backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
google-generativeai>=0.3.0
mem0ai>=0.0.12
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
websockets>=11.0
